[feat] 호스팅한 펀딩 리스트 완료 api 연결

+++ app/src/main/java/com/example/habday_android/src/main/list/detail/delete/DeleteFundingDialog.kt
@@ -10,7 +10,7 @@ import com.example.habday_android.databinding.DialogDeleteFundingBinding
 
 class DeleteFundingDialog(context: Context): Dialog(context) {
     private lateinit var binding: DialogDeleteFundingBinding
-    var itemId : Long? = null
+    var itemId : Int? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
